server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.118.0",
    "httpx>=0.28.1",
    "logfire[fastapi,system-metrics]>=4.14.2",
    "pydantic-settings>=2.11.0",
    "python-jose>=3.5.0",
    "redis>=6.4.0",
    "uvicorn[standard]>=0.37.0",
]
